chore: move tiny tests in it own file
